+++ src/jsontree.js
@@ -4,7 +4,7 @@
  * A lightweight JavaScript library that generates customizable tree views to better visualize JSON data.
  * 
  * @file        jsontree.js
- * @version     v0.1.0
+ * @version     v0.2.0
  * @author      Bunoon
  * @license     MIT License
  * @copyright   Bunoon 2024
@@ -202,6 +202,10 @@
 
         if ( bindingOptions.sortPropertyNames ) {
             properties = properties.sort();
+
+            if ( !bindingOptions.sortPropertyNamesInAlphabeticalOrder ) {
+                properties = properties.reverse();
+            }
         }
 
         var propertiesLength = properties.length;
@@ -210,7 +214,7 @@
             var propertyName = properties[ propertyIndex ];
 
             if ( data.hasOwnProperty( propertyName ) ) {
-                renderValue( objectTypeContents, bindingOptions, propertyName, data[ propertyName ] );
+                renderValue( objectTypeContents, bindingOptions, propertyName, data[ propertyName ], propertyIndex === propertiesLength - 1 );
                 propertyCount++;
             }
         }
@@ -226,13 +230,13 @@
         for ( var dataIndex = 0; dataIndex < dataLength; dataIndex++ ) {
             var name = bindingOptions.useZeroIndexingForArrays ? dataIndex.toString() : ( dataIndex + 1 ).toString();
 
-            renderValue( objectTypeContents, bindingOptions, name, data[ dataIndex ] );
+            renderValue( objectTypeContents, bindingOptions, name, data[ dataIndex ], dataIndex === dataLength - 1 );
         }
 
         addArrowEvent( bindingOptions, arrow, objectTypeContents );
     }
 
-    function renderValue( container, bindingOptions, name, value ) {
+    function renderValue( container, bindingOptions, name, value, isLastItem ) {
         var objectTypeValue = createElement( container, "div", "object-type-value" ),
             arrow = bindingOptions.showArrowToggles ? createElement( objectTypeValue, "div", "no-arrow" ) : null,
             valueElement = null;
@@ -243,21 +247,32 @@
         if ( !isDefined( value ) ) {
             valueElement = createElementWithHTML( objectTypeValue, "span", "null", "null" );
 
+            createComma( bindingOptions, objectTypeValue, isLastItem );
+
         } else if ( isDefinedFunction( value ) ) {
             valueElement = createElementWithHTML( objectTypeValue, "span", "function", getFunctionName( value ) );
+            
+            createComma( bindingOptions, objectTypeValue, isLastItem );
 
         }  else if ( isDefinedBoolean( value ) ) {
             valueElement = createElementWithHTML( objectTypeValue, "span", "boolean", value );
+            
+            createComma( bindingOptions, objectTypeValue, isLastItem );
 
         } else if ( isDefinedNumber( value ) ) {
             valueElement = createElementWithHTML( objectTypeValue, "span", "number", value );
+            
+            createComma( bindingOptions, objectTypeValue, isLastItem );
 
         } else if ( isDefinedString( value ) ) {
             valueElement = createElementWithHTML( objectTypeValue, "span", "string", bindingOptions.showStringQuotes ? "\"" + value + "\"" : value );
+            
+            createComma( bindingOptions, objectTypeValue, isLastItem );
 
         } else if ( isDefinedDate( value ) ) {
             valueElement = createElementWithHTML( objectTypeValue, "span", "date", getCustomFormattedDateTimeText( value, bindingOptions.dateTimeFormat ) );
-            
+            createComma( bindingOptions, objectTypeValue, isLastItem );
+
         } else if ( isDefinedObject( value ) && !isDefinedArray( value ) ) {
             var objectTitle = createElement( objectTypeValue, "span", "object" ),
                 objectTypeContents = createElement( objectTypeValue, "div", "object-type-contents" ),
@@ -269,6 +284,8 @@
                 createElementWithHTML( objectTitle, "span", "count", "{" + propertyCount + "}" );
             }
 
+            createComma( bindingOptions, objectTitle, isLastItem );
+
         } else if ( isDefinedArray( value ) ) {
             var arrayTitle = createElement( objectTypeValue, "span", "array" ),
                 arrayTypeContents = createElement( objectTypeValue, "div", "object-type-contents" );
@@ -279,6 +296,7 @@
                 createElementWithHTML( arrayTitle, "span", "count", "[" + value.length + "]" );
             }
 
+            createComma( bindingOptions, arrayTitle, isLastItem );
             renderArrayValues( arrow, arrayTypeContents, bindingOptions, value );
         }
 
@@ -335,6 +353,12 @@
         return result;
     }
 
+    function createComma( bindingOptions, objectTypeValue, isLastItem ) {
+        if ( bindingOptions.showCommas && !isLastItem ) {
+            createElementWithHTML( objectTypeValue, "span", "comma", "," );
+        }
+    }
+
 
     /*
      * ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
@@ -354,6 +378,8 @@
         options.showTitleTreeControls = getDefaultBoolean( options.showTitleTreeControls, true );
         options.showAllAsClosed = getDefaultBoolean( options.showAllAsClosed, false );
         options.sortPropertyNames = getDefaultBoolean( options.sortPropertyNames, true );
+        options.sortPropertyNamesInAlphabeticalOrder = getDefaultBoolean( options.sortPropertyNamesInAlphabeticalOrder, true );
+        options.showCommas = getDefaultBoolean( options.showCommas, false );
 
         options = buildAttributeOptionStrings( options );
         options = buildAttributeOptionCustomTriggers( options );
@@ -609,6 +635,48 @@
 
     /*
      * ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+     * Public Functions:  Manage Instances
+     * ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+     */
+
+    /**
+     * render().
+     * 
+     * Renders an element using the options specified.
+     * 
+     * @public
+     * 
+     * @param       {Object}    element                                     The element to render.
+     * @param       {Object}    options                                     The options to use (refer to "Binding Options" documentation for properties).
+     * 
+     * @returns     {Object}                                                The JsonTree.js class instance.
+     */
+    this.render = function( element, options ) {
+        if ( isDefinedObject( element ) && isDefinedObject( options ) ) {
+            renderControl( renderBindingOptions( options, element ) );
+        }
+
+        return this;
+    };
+
+    /**
+     * renderAll().
+     * 
+     * Finds all new elements and renders them.
+     * 
+     * @public
+     * 
+     * @returns     {Object}                                                The JsonTree.js class instance.
+     */
+    this.renderAll = function() {
+        render();
+
+        return this;
+    };
+
+
+    /*
+     * ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
      * Public Functions:  Configuration
      * ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
      */
@@ -668,7 +736,7 @@
      * @returns     {string}                                                The version number.
      */
     this.getVersion = function() {
-        return "0.1.0";
+        return "0.2.0";
     };